Create nn folder
