+++ db_transformer/nn/layers/__init__.py
@@ -1,3 +1,2 @@
-from .hetero_gnn import HeteroGNNLayer
 from .node_applied import NodeApplied
 from .per_feature_norm import PerFeatureNorm
